requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.8
-e git+https://github.com/argoai/argoverse-api.git@f886ac54fba9f06f8a7d109eb663c7f501b3aa8e#egg=argoverse
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
cachetools==5.3.0
charset-normalizer==3.1.0
colour==0.1.5
cycler==0.11.0
descartes==1.1.0
easydict==1.10
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
glob2==0.7
google-auth==2.18.0
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
grpcio==1.54.0
h5py==3.8.0
hydra-core==1.1.0
imageio==2.27.0
imgaug==0.4.0
importlib-metadata==4.13.0
importlib-resources==5.12.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lapsolver==1.1.0
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
motmetrics==1.1.3
multidict==6.0.4
networkx==2.6.3
numba==0.56.4
numpy==1.19.0
oauthlib==3.2.2
omegaconf==2.1.0
opencv-python==4.7.0.72
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
polars==0.17.12
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
pyntcloud==0.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-louvain==0.16
pytorch-lightning==1.5.2
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
rdflib==6.3.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
shapely==2.0.1
spconv==1.2.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.8.0+cu111
torch-cluster==1.6.0
torch-geometric==1.7.2
torch-scatter==2.1.1
torch-sparse==0.6.10
torch-spline-conv==1.2.2
torchaudio==0.8.0
torchmetrics==0.11.4
torchvision==0.9.0+cu111
tqdm==4.65.0
typing_extensions==4.5.0
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.15.0
